Extract TextSuperFormField tests and improve docs
